// src/lib/types.ts
export interface Invoice {
  invoice: string;
  payment_hash: string;
  expires_in: number;
  donor_name?: string;
  amount_sats: number;
}

export interface InvoiceStatus {
  status: "PENDING" | "PAID";
  paid_at?: string;
}

export interface Donation {
  id: string;
  donor_name: string;
  amount_sats: number;
  status: "PAID";
  paid_at: string;
}